[doc] src module name changed to pypipe
