+++ pypipe/core/interfaces.py
