+++ pypipe/core/paths.py
@@ -1,0 +1,14 @@
+PREPROCESSING_CONFIG_PATH = "pypipe/configs/preprocessing.json"
+
+
+MODEL_DEFAULT_PATHS = {
+    "countvec": "pypipe/data/models/countvec",
+    "word2vec": "pypipe/data/models/word2vec",
+    "regex_norm": "pypipe/data/models/regex_norm"
+}
+
+
+VOCAB_DEFAULT_PATHS = {
+    "countvec": "pypipe/data/corpus/countvec",
+    "word2vec": "pypipe/data/corpus/word2vec"
+}
