+++ pypipe/data/corpus/corpus.txt
@@ -1,0 +1,4 @@
+HOLA!!1111 gente lindaaaaa!!!
+el nombre essssss @Pedro re loco jjajajajjja
+mi correo es pedrito@gmail.com
+su página es www.pedrito.com ....
