+++ pypipe/data/corpus/new_corpus.txt
@@ -1,0 +1,3 @@
+hola gente
+jaja su nombre es pedro
+el correo de pedro es MAIL
