+++ pypipe/data/models/pruebas/vocab.json
@@ -1,0 +1,18 @@
+{
+    "hola": 5,
+    "gente": 4,
+    "linda": 7,
+    "el": 1,
+    "nombre": 12,
+    "es": 3,
+    "mention": 9,
+    "muy": 11,
+    "loco": 8,
+    "jaja": 6,
+    "mi": 10,
+    "correo": 0,
+    "email": 2,
+    "su": 14,
+    "pagina": 13,
+    "url": 15
+}
