+++ pypipe/data/models/regex_norm/normcorpus.txt
@@ -1,0 +1,4 @@
+hola gente linda
+el nombre es mention muy loco jaja
+mi correo es email
+su pagina es url
