+++ pypipe/data/models/word2vec/vocab.json
@@ -1,0 +1,4 @@
+{
+    "hola": 0,
+    "chau": 1
+}
