--- src/core/paths.py
@@ -1,14 +1,0 @@
-PREPROCESSING_CONFIG_PATH = "src/configs/preprocessing.json"
-
-
-MODEL_DEFAULT_PATHS = {
-    "countvec": "src/data/models/countvec",
-    "word2vec": "src/data/models/word2vec",
-    "regex_norm": "src/data/models/regex_norm"
-}
-
-
-VOCAB_DEFAULT_PATHS = {
-    "countvec": "src/data/corpus/countvec",
-    "word2vec": "src/data/corpus/word2vec"
-}
